% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergev-package.R
\docType{package}
\name{mergev-package}
\alias{mergev-package}
\title{Gives verbose feedback on merges, including the merge type (one/many to one/many), information
on missing key variables, and a merge statistics table.}
\description{
See \code{\link{mergev}} for examples.
}

